Change contact view by merging local contacts and direct chats Change room view by renaming french version from salon to discussion, and adding direct chats in the list Change several base colors Little improving in search people view More controls for external users
