+++ vector/src/main/res/values/strings.xml
@@ -330,7 +330,7 @@
     <string name="room_title_one_member">1 member</string>
     <string name="room_e2e_alert_title">Warning!</string>
     <string name="room_e2e_alert_message">End-to-end encryption is in beta and may not be reliable.\n\nYou should not yet trust it to secure data.\n\nDevices will not yet be able to decrypt history from before they joined the room.\n\nEncrypted messages will not be visible on clients that do not yet implement encryption.</string>
-
+    <string name="room_creation_forbidden">You are not authorized to create a room</string>
     <!--  Time -->
     <string name="format_time_s">s</string>
     <string name="format_time_m">m</string>
